rebase: has been rebase with the main branch

+++ src/view/battle/console/BattlePokemonConsole.java
@@ -71,7 +71,7 @@ public class BattlePokemonConsole implements ViewBattle {
             int indexPokemon2 = choosePokemon(controller.trainer2);
             controller.startCombat(indexPokemon1, indexPokemon2);
             do {
-                int index = (controller.getTurn()) ? selectAttack(controller.getPokemon1()): selectAttack(controller.getPokemon2());
+                byte index = (controller.getTurn()) ? selectAttack(controller.getPokemon1()): selectAttack(controller.getPokemon2());
                 controller.processAttack(index);
             }while(!controller.hasFinish());
         }while(!(BattleTrainer.trainerHasLost(controller.trainer1) ||  BattleTrainer.trainerHasLost(controller.trainer2)));
@@ -88,16 +88,16 @@ public class BattlePokemonConsole implements ViewBattle {
             System.out.println(controller.getHistory().ShowAllPila());}
     }
 
-    public int selectAttack(Pokemon pokemon) {
+    public byte selectAttack(Pokemon pokemon) {
         Scanner sc = new Scanner(System.in);
         boolean condition = false;
-        int index = -1;
+        byte index = -1;
         System.out.printf(Colors.WHITE_BOLD + "\nTurno de %s%n" + Colors.RESET, controller.getTurn() ? controller.trainer1.getNameTrainer() : controller.trainer2.getNameTrainer());
         do {
             PokemonMenu.showPokemonAttacks(pokemon);
 
             try{
-                index = Integer.parseInt(sc.nextLine())-1;
+                index =(byte)(Integer.parseInt(sc.nextLine())-1);
                 condition = !(index >= 0 && index < pokemon.getAttacks().length);
                 if(condition){
                     throw new InvalidAttackSelectionException("⚠️Por favor ingresa uno de los ataques disponibles!");
